pyproject.toml
[tool.poetry]
name = "codename-score-backend"
version = "0.1.0"
description = ""
authors = ["beom-yun <scds0326@naver.com>"]
readme = "README.md"
packages = [{include = "codename_score_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.3"
black = "^22.10.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
